requirements.txt
requests==2.31.0
newsapi-python==0.2.7
threads-sdk==0.2.2 